+++ changelog.md
@@ -17,6 +17,11 @@ r"^\#\# \[\d{1,}[.]\d{1,}[.]\d{1,}\] \- \d{4}\-\d{2}-\d{2}$"
 -->
 
 ## Released
+## [0.9.0] - 2022-07-30
+### Added
+- Support GPIO usage with [`nextion_gpio`](nextion/nextion_gpio.py)
+- [Usage example](examples/gpio/main.py) of GPIO
+
 ## [0.8.0] - 2022-07-30
 ### Added
 - Support all class specific functions of
@@ -137,8 +142,10 @@ r"^\#\# \[\d{1,}[.]\d{1,}[.]\d{1,}\] \- \d{4}\-\d{2}-\d{2}$"
 - [Example HMI file](examples/everything.HMI) to be used for all examples
 
 <!-- Links -->
-[Unreleased]: https://github.com/brainelectronics/micropython-nextion/compare/0.7.1...develop
+[Unreleased]: https://github.com/brainelectronics/micropython-nextion/compare/0.9.0...develop
 
+[0.9.0]: https://github.com/brainelectronics/micropython-nextion/tree/0.9.0
+[0.8.0]: https://github.com/brainelectronics/micropython-nextion/tree/0.8.0
 [0.7.1]: https://github.com/brainelectronics/micropython-nextion/tree/0.7.1
 [0.7.0]: https://github.com/brainelectronics/micropython-nextion/tree/0.7.0
 [0.6.0]: https://github.com/brainelectronics/micropython-nextion/tree/0.6.0